requirements.txt
flask
getmac
PyJWT
cryptography
flask_sqlalchemy
python-dotenv